migrate rmf_panel js component to a new repo
