--- .github/workflows/gh_page.yml
@@ -1,30 +1,0 @@
-name: build-panel
-on:
-  push:
-    branches: [ main ]
-  pull_request:
-    branches: [ main ]
-jobs:
-  build-and-deploy:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          path: rmf_demos
-      - uses: actions/setup-node@v2
-        with:
-          node-version: '14'
-          check-latest: true
-      - name: build-panel
-        run: |
-          cd rmf_demos/rmf_demos_panel/rmf_demos_panel/static
-          npm install
-          npm run build
-      # This will run only during merge to main
-      - name: deploy_page
-        if: github.event_name == 'push'
-        uses: peaceiris/actions-gh-pages@v3
-        with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
-          publish_dir: rmf_demos/rmf_demos_panel/rmf_demos_panel/static/dist
-          destination_dir: v1.0.0 # Change the version of webpack: "app.bundle.js"
