+++ rmf_demos/launch/dashboard.launch.xml
@@ -3,21 +3,18 @@
 
 <launch>
   <arg name="use_sim_time" default="true" description="Use the /clock topic for time to sync with simulation"/>
-  <arg name="server_ip" default="0.0.0.0" description="GUI IP address"/>
+  <arg name="server_ip" default="0.0.0.0" description="RMF_demos simple api-server IP address"/>
+  <arg name="server_port_num" default="8083" description="RMF_demos simple api-server Port number"/>
   <arg name="dashboard_config_file" default="" description="Path to dashboard config for web rmf panel file"/>
 
   <!-- Dispatcher API Server -->
+  <!-- Note:  not to be confused with the `simple-api-server` in rmf_demos 
+              and "prod version" of `api-server` in `rmf-web` -->
   <group>
-    <node pkg="rmf_demos_panel" exec="api_server"  output="screen">
+    <node pkg="rmf_demos_panel" exec="simple_api_server"  output="screen">
       <param name="use_sim_time" value="$(var use_sim_time)"/>
-      <env name="WEB_SERVER_IP_ADDRESS" value="$(var server_ip)" />
-    </node>
-  </group>
-
-  <!-- Dashboard GUI Server -->
-  <group>
-    <node pkg="rmf_demos_panel" exec="gui_server"  output="screen">
-      <env name="WEB_SERVER_IP_ADDRESS" value="$(var server_ip)" />
+      <env name="RMF_DEMOS_API_SERVER_IP" value="$(var server_ip)" />
+      <env name="RMF_DEMOS_API_SERVER_PORT" value="$(var server_port_num)" />
       <env name="DASHBOARD_CONFIG_PATH" value="$(var dashboard_config_file)" />
     </node>
   </group>
