--- rmf_demos_panel/rmf_demos_panel/static/.gitignore
@@ -1,5 +1,0 @@
-node_modules
-dist
-web_server.log
-package-lock.json
-coverage
