--- rmf_demos_panel/rmf_demos_panel/static/package.json
@@ -1,61 +1,0 @@
-{
-  "name": "rmf_demos_panel",
-  "version": "1.0.0",
-  "description": "",
-  "main": "main.js",
-  "scripts": {
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:cov": "jest --coverage",
-    "build": "webpack --progress --config webpack.config.js",
-    "build-dev": "webpack --mode development --progress --config webpack.config.js",
-    "dev-build": "webpack --progress --config webpack.config.js",
-    "watch": "webpack --progress --config webpack.config.js --watch"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/osrf/rmf_demos.git"
-  },
-  "author": "OSRF",
-  "babel": {
-    "presets": [
-      "@babel/preset-env",
-      "@babel/preset-react"
-    ]
-  },
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/osrf/rmf_demos/issues"
-  },
-  "dependencies": {
-    "@material-ui/core": "^4.11.1",
-    "@material-ui/icons": "^4.11.2",
-    "antd": "^4.13.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "socket.io-client": "^3.0.4",
-    "typescript": "^4.1.2"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.12.9",
-    "@babel/preset-env": "^7.12.7",
-    "@babel/preset-react": "^7.12.7",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@testing-library/jest-dom": "^5.11.6",
-    "@testing-library/react": "^11.2.2",
-    "@testing-library/user-event": "^12.6.0",
-    "@types/jest": "^26.0.19",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/socket.io-client": "^1.4.34",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "^8.2.1",
-    "file-loader": "^6.2.0",
-    "jest": "^26.6.3",
-    "json-loader": "^0.5.7",
-    "ts-jest": "^26.4.4",
-    "ts-loader": "^8.0.11",
-    "webpack": "^5.10.0",
-    "webpack-cli": "^4.2.0"
-  }
-}
