--- rmf_demos_panel/rmf_demos_panel/static/setupTests.js
@@ -1,1 +1,0 @@
-import '@testing-library/jest-dom/extend-expect';
