--- rmf_demos_panel/rmf_demos_panel/static/src/__tests__/footer.test.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import Footer from '../components/fixed-components/footer';
-
-describe('Footer', () => {
-    it("should render", () => {
-        render(<Footer />);
-        expect(screen.getByRole('footer')).toBeVisible();
-    
-    });
-})
