--- rmf_demos_panel/rmf_demos_panel/static/src/__tests__/header.test.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { render, screen } from '@testing-library/react';
-import Header from '../components/fixed-components/header';
-
-describe('Header', () => {
-    it("should render", () => {
-        render(<Header />);
-        expect(screen.getByRole('header')).toBeVisible();
-    
-    });
-})
