--- rmf_demos_panel/rmf_demos_panel/static/src/components/socket.tsx
@@ -1,5 +1,0 @@
-import io from 'socket.io-client';
-
-const ENDPOINT = "http://" + location.hostname + ":8080/status_updates"
-
-export const socket = io.connect(ENDPOINT);
