--- rmf_demos_panel/rmf_demos_panel/static/src/index.tsx
@@ -1,5 +1,0 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
-import App from './app';
-
-ReactDOM.render(<App />, document.getElementById("root"));
