--- rmf_demos_panel/rmf_demos_panel/static/tsconfig.json
@@ -1,17 +1,0 @@
-{
-    "compilerOptions": {
-        "moduleResolution": "Node",
-        "sourceMap": true,
-        "module": "commonjs",
-        "target": "es5",
-        "jsx": "react",
-        "resolveJsonModule": true,
-        "esModuleInterop": true,
-        "allowJs": true
-    },
-    "include": [
-        "./src/**/*.tsx",
-        "*.d.ts"
-    ],
-    "exclude": ["./node_modules"]
-}
