--- rmf_demos_panel/rmf_demos_panel/templates/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" dir="ltr">
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/3.0.5/socket.io.js"></script>
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css">
-<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/antd/4.9.4/antd.min.css">
-<link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
-<head>
-  <title>RMF Panel</title>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-
-<body>
-  <div id="root">
-    <p>Loading... static/dist/app.bundle.js.</p>
-    </p>
-  </div>
-<hr>
-
-<script src="../static/dist/app.bundle.js" type="text/javascript"></script>
-</body>
-</html>
